requirements.txt
python-dotenv
google-genai
streamlit
chromadb
pypdf